requirements.txt
pytest
requests
allure-pytest
pyYaml